+++ lib/consts/paths.const.sh
@@ -1,1 +1,2 @@
 BASE_PATH=/home/admin
+SCRIPTS_PATH=/home/.scripts