+++ lib/modules/coretools/install_coretools.sh
@@ -1,0 +1,5 @@
+install_coretools () {
+    if ! $(is_scripts_installed) -eq "true"; then
+        install_scripts
+    fi
+}