+++ lib/modules/coretools/verify_coretool.sh
@@ -1,0 +1,10 @@
+verify_coretool () {
+    local service=$2
+    local condition=$1
+    
+    if "$condition" -eq "true"; then
+        print "  ${SYMBOL_OK} $service"
+    else
+        print "  ${SYMBOL_ERROR} $service"
+    fi
+}