+++ lib/modules/coretools/verify_coretools.sh
@@ -1,0 +1,5 @@
+verify_services () {
+    print "Coretools:"
+    
+    verify_coretool "$(is_scripts_installed)" "Bash Scripts"
+}